+++ spec/shelly/cli/backup_spec.rb
@@ -233,7 +233,7 @@ describe Shelly::CLI::Backup do
     end
 
     it "should restore database" do
-      $stdout.should_receive(:puts).with("You are about restore postgresql database for cloud foo-staging to state from better.tar.gz")
+      $stdout.should_receive(:puts).with("You are about to restore postgresql database for cloud foo-staging to state from better.tar.gz")
       $stdout.should_receive(:print).with("I want to restore the database (yes/no): ")
       $stdout.should_receive(:puts).with("\n")
       @client.stub(:restore_backup).with("todo-list-test","better.tar.gz")
@@ -247,7 +247,7 @@ describe Shelly::CLI::Backup do
 
     context "when answering no" do
       it "should cancel restore database" do
-        $stdout.should_receive(:puts).with("You are about restore postgresql database for cloud foo-staging to state from better.tar.gz")
+        $stdout.should_receive(:puts).with("You are about to restore postgresql database for cloud foo-staging to state from better.tar.gz")
         $stdout.should_receive(:print).with("I want to restore the database (yes/no): ")
         $stdout.should_receive(:puts).with("\n")
         $stdout.should_receive(:puts).with(red "Canceled")